+++ templates/MicroserviceSourceGen/MicroserviceSourceGen/Diagnostics.cs
@@ -56,6 +56,9 @@ public static class Diagnostics
 		public const string CALLABLE_METHOD_TYPE_INSIDE_MICROSERVICE_SCOPE_ID = "BEAM_SRV_O007";
 		public const string CALLABLE_METHOD_TYPE_IS_NESTED_ID = "BEAM_SRV_O008";
 		public const string CLASS_BEAM_GENERATE_SCHEMA_ATTRIBUTE_IS_NESTED_ID = "BEAM_SRV_O009";
+		public const string MICROSERVICE_ID_INVALID_FROM_CS_PROJ_ID = "BEAM_SRV_0010";
+
+		public const string PROP_BEAM_ID = "BeamId";
 
 		public static readonly DiagnosticDescriptor NoMicroserviceClassesDetected
 			= new(NO_MICROSERVICE_DETECTED_DIAGNOSTIC_ID,
@@ -134,6 +137,15 @@ public static class Diagnostics
 				Category_Services,
 				DiagnosticSeverity.Error,
 				helpLinkUri: "https://docs.beamable.com/docs/cli-guide-microservices#beam-generated-schema-class-is-a-nested-type",
+				isEnabledByDefault: true);
+		
+		public static readonly DiagnosticDescriptor MicroserviceIdInvalidFromCsProj
+			= new(MICROSERVICE_ID_INVALID_FROM_CS_PROJ_ID,
+				$"{nameof(Server.Microservice)} ID is invalid, it needs to be the same as <BeamId> csharp property (or as csproj name if none exists)",
+				$"{nameof(Server.Microservice)} ID: `{{0}}` is invalid, it needs to be the same as <BeamId> csharp property (or as csproj name if none exists): `{{1}}`",
+				Category_Services,
+				DiagnosticSeverity.Error,
+				helpLinkUri: "https://docs.beamable.com/docs/cli-guide-microservices#invalid-microservice-id",
 				isEnabledByDefault: true);
 	}